+++ events_boilerplate.go
@@ -335,6 +335,14 @@ func (e *EventUnlinkGithubType) TypeDataPtr() interface{} {
 	return &e.TypeData
 }
 
+func (e *EventDeleteProjectType) TypeDataPtr() interface{} {
+	return &e.TypeData
+}
+
 func (e *EventNewProjectType) TypeDataPtr() interface{} {
+	return &e.TypeData
+}
+
+func (e *EventEditProjectType) TypeDataPtr() interface{} {
 	return &e.TypeData
 }