+++ events_specialize.go
@@ -178,8 +178,12 @@ func (pev *Event) Specialize() DetailedEvent {
 		e = &EventLinkGithubType{Event: ev}
 	case EventUnlinkGithub:
 		e = &EventUnlinkGithubType{Event: ev}
+	case EventDeleteProject:
+		e = &EventDeleteProjectType{Event: ev}
 	case EventNewProject:
 		e = &EventNewProjectType{Event: ev}
+	case EventEditProject:
+		e = &EventEditProjectType{Event: ev}
 	default:
 		return pev
 	}