+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## To Be Released
 
+* feat(projects): add edit project event
+* feat(projects): add delete project event
 * feat(projects): add new project created event
 * feat(projects): add update app project event